+++ src/CMakeLists.txt
@@ -1,7 +1,7 @@
 include_directories(.)
 
 add_subdirectory(Common)
-add_subdirectory(Renderer)
+add_subdirectory(Video)
 add_subdirectory(Engine)
 add_subdirectory(Editor)
 add_subdirectory(Game)
