+++ src/Engine/CMakeLists.txt
@@ -191,6 +191,6 @@ endforeach()
 set_property(SOURCE ${SRCS} APPEND PROPERTY OBJECT_DEPENDS ${EMBEDDED_HEADER})
 
 add_library(Engine STATIC ${SRCS} ${HEADERS} ${EMBEDDED_SOURCE})
-target_link_libraries(Engine Renderer imgui stb simpleini jsoncpp_lib_static ${OPENAL_LIB} angelscript assimp)
+target_link_libraries(Engine Video imgui stb simpleini jsoncpp_lib_static ${OPENAL_LIB} angelscript assimp)
 set_property(TARGET Engine PROPERTY CXX_STANDARD 11)
 set_property(TARGET Engine PROPERTY CXX_STANDARD_REQUIRED ON)
