+++ src/Engine/Manager/DebugDrawingManager.cpp
@@ -7,8 +7,8 @@
 #include "../MainWindow.hpp"
 #include "Managers.hpp"
 #include "ResourceManager.hpp"
-#include <Renderer/Shader/Shader.hpp>
-#include <Renderer/Shader/ShaderProgram.hpp>
+#include <Video/Shader/Shader.hpp>
+#include <Video/Shader/ShaderProgram.hpp>
 #include "DebugDrawing.vert.hpp"
 #include "DebugDrawing.frag.hpp"
