+++ src/Engine/Manager/ParticleManager.cpp
@@ -9,8 +9,8 @@
 #include "../MainWindow.hpp"
 #include "../Manager/Managers.hpp"
 #include "../Manager/ResourceManager.hpp"
-#include <Renderer/Shader/Shader.hpp>
-#include <Renderer/Shader/ShaderProgram.hpp>
+#include <Video/Shader/Shader.hpp>
+#include <Video/Shader/ShaderProgram.hpp>
 #include "../Texture/Texture2D.hpp"
 #include "Particle.vert.hpp"
 #include "Particle.geom.hpp"
