+++ src/Engine/RenderTarget.cpp
@@ -1,8 +1,8 @@
 #include "RenderTarget.hpp"
 #include <Common/Log.hpp>
 
-#include <Renderer/Shader/Shader.hpp>
-#include <Renderer/Shader/ShaderProgram.hpp>
+#include <Video/Shader/Shader.hpp>
+#include <Video/Shader/ShaderProgram.hpp>
 
 #include "Geometry/Rectangle.hpp"
